Update BasicIO_Terms.md
Asynchronous/Synchronous - These two terms talk about a “shared clock”; with no context on what a shared clock is, my first impression is to either define “shared clock” or better define what Asynchronous/Synchronous mean. For example, dictate why they come by their names, like in this article: https://www.techwalla.com/articles/difference-between-synchronous-and-asynchronous-data-transfer. You might also want to provide an example or two for each type: like audio/video chats for synchronous and emails/forums. I know you probably talk more about terms on within the topic you introduce them, but consider that you are providing these terms on other pages in which they are not discussed. You told me the other day, you don’t want them to have to go hunting on another page for the full explanation (aka, what does that really mean?). In some cases, yeah, they will have to re-read to understand the concept, but the term still should be a complete and thoughtful nutshell-like buoy in an ocean of concepts. Anyway, I’ll let you think on that.

Will the students know what a CPU is? Is that why it’s italicized, because it’s defined somewhere else? I’d definitely spell it out the first time it’s introduced, as you can see in my edit.

The terms “interrupt” and “polling” start strangely, like in the middle of a sentence. Try re-organizing them to be clearer.

I like how you defined “Priorities”. You classify it as a system that does X. This style is clear and it would nice to see other definitions mirror it in some way. You’ll see I attempted to do that with most of the other definitions (other than “interrupt” and “polling” because I’m not sure what they are classified as).